Move NetworkIDsToName and other functions to new file + change it to IntEnum
